+++ src/main/scala/com/chinawiserv/wsmp/occupancy/model/OccupancyData.scala
@@ -5,4 +5,4 @@ import scala.collection.mutable.ArrayBuffer
 /**
   * Created by zengpzh on 2017/1/11.
   */
-private[occupancy] case class OccupancyData(id: Int, time: String, levels: ArrayBuffer[Short])
+private[occupancy] case class OccupancyData(id: Int, time: String, levels: ArrayBuffer[Byte])
